token extractor + user extractor

+++ part4/bloglist/controllers/blogcontrollers.js
@@ -2,7 +2,8 @@ const blogRouter = require('express').Router()
 const Blog = require("../models/blog.js");
 require('express-async-errors')
 const User = require('../models/user') 
-
+const jwt = require('jsonwebtoken')
+const tokenExtractor = require('../middleware/tokenExtract');
 
 
   blogRouter.get('/blogs' , async (req , res) => {
@@ -10,10 +11,13 @@ const User = require('../models/user')
     return res.json(blogs)
     
   })
+
+
   
-  blogRouter.post('/blogs', async (request, response) => {
+  blogRouter.post('/blogs', tokenExtractor, async (request, response) => {
     const blog = new Blog(request.body)
-    const user = await User.findById(request.body.userId)
+    const user = await User.findById(request.user.id)
+   
     blog.user = user.id
 
     if(!blog.likes){
@@ -25,7 +29,21 @@ const User = require('../models/user')
     response.status(201).json(savedBlog)
   })
 
-  blogRouter.delete('/blogs/:id', async (req,res) =>{
+  blogRouter.delete('/blogs/:id',tokenExtractor, async (req,res) =>{
+      const blog = await Blog.findById(req.params.id);
+      const user = req.user
+      if(!blog){
+        return res.status(404).json({error:'blog not found'})
+      }
+      if(!user){
+        return res.status(401).json({ error: 'you have to be logged in' })
+      }
+      if(!blog.user){
+        return res.status(401).json({error:'no token was found'})
+      }
+      if(user.id !== blog.user.toString()) {
+        return res.status(403).json({ error: 'You cant delete other users posts' })
+      }
       await Blog.findByIdAndDelete(req.params.id)
       res.status(204).end()
   })
